remove src/opentrons/main.py and move the code into src/opentrons/__init__.py
